Animal Bugfix, New seek method with 2 parameters
cohesion now works as follows: 2 radius, inner and outer radius. ratio variable computes the ratio between the inner and outer radius. The ratio is given to the new seek parameter to modify the amplitude of the resulting vector. This means animals in the outer radius influnce the movement less. -> less wobbly